flyttar ut voting-actions till egen actions-fil
